tickets/src/events/listener/__test__/order-cancelled-listener.test.ts
import { OrderCancelledEvent } from "@wsticketing/common";
import { Message } from "node-nats-streaming";
import mongoose from "mongoose";

import { OrderCancelledListener } from "../order-cancelled-listener";
import { natsWrapper } from "../../../nats-wrapper";
import { buildTicket } from "../../../test/setup";
import { Ticket } from "../../../models/ticket";

const setup = async () => {
  const listener = new OrderCancelledListener(natsWrapper.client);

  const orderId = new mongoose.Types.ObjectId().toHexString();
  const ticket = await buildTicket();
  ticket.set({
    orderId,
  });
  await ticket.save();

  const data: OrderCancelledEvent["data"] = {
    id: ticket.orderId!,
    version: ticket.version,
    ticket: {
      id: ticket.id,
    },
  };

  // @ts-ignore
  const msg: Message = {
    ack: jest.fn(),
  };

  return { listener, ticket, data, msg };
};

it("should reset the orderId", async () => {
  const { listener, ticket, data, msg } = await setup();

  expect(ticket.orderId).toEqual(data.id);

  await listener.onMessage(data, msg);

  const updatedTicket = await Ticket.findById(ticket.id);

  expect(updatedTicket!.orderId).not.toBeDefined();
});

it("acks the message", async () => {
  const { listener, data, msg } = await setup();

  await listener.onMessage(data, msg);

  expect(msg.ack).toHaveBeenCalled();
});

it("emits a ticket-updated-event", async () => {
  const { listener, data, msg } = await setup();

  await listener.onMessage(data, msg);

  expect(natsWrapper.client.publish).toHaveBeenCalled();

  const ticketUpdatedData = JSON.parse(
    (natsWrapper.client.publish as jest.Mock).mock.calls[0][1]
  );

  expect(ticketUpdatedData.orderId).not.toBeDefined();
});
